change config

+++ normuon.py
@@ -87,7 +87,8 @@ def _muon_step_impl(stacked_grads, stacked_params,
     v_norm_sq = v_mean.sum(dim=(-2, -1), keepdim=True) * red_dim_size
     v_norm = v_norm_sq.sqrt()
     second_momentum_buffer.lerp_(
-        v_mean.to(dtype=second_momentum_buffer.dtype), 1 - beta2
+        v_mean.to(dtype=second_momentum_buffer.dtype),
+        (1 - beta2).to(dtype=second_momentum_buffer.dtype),
     )
     step_size = second_momentum_buffer.clamp_min(1e-10).rsqrt()
     scaled_sq_sum = (v_mean * red_dim_size) * step_size.float().square()
